+++ SUDATalk/Network/Test/URLSessionProtocol.swift
@@ -1,0 +1,14 @@
+//
+//  URLSessionProtocol.swift
+//  SUDATalk
+//
+//  Created by 박다현 on 1/2/25.
+//
+
+import Foundation
+
+protocol URLSessionProtocol {
+    func dataTask(with request: URLRequest, completionHandler: @escaping @Sendable (Data?, URLResponse?, (any Error)?) -> Void) -> URLSessionDataTask
+}
+
+extension URLSession: URLSessionProtocol {}
